requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.23.9
botocore==1.26.10
celery==5.2.6
celery-progress==0.1.2
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.6
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.4
django-celery-results==2.3.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-restframework==0.0.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.0
kombu==5.2.4
MarkupSafe==2.1.1
mysqlclient==2.1.0
oauthlib==3.2.0
packaging==21.3
prompt-toolkit==3.0.29
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
redis==4.3.1
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
flower
m3u8
ffmpeg-python
